Get users profiles by specific conditions

+++ app/src/main/java/com/krygodev/singlesplanet/profile/ProfileScreen.kt
@@ -280,8 +280,8 @@ fun ProfileScreen(
                             viewModel.onEvent(
                                 ProfileEvent.UpdateSearchAge(
                                     listOf(
-                                        it.start,
-                                        it.endInclusive
+                                        it.start.toInt().toFloat(),
+                                        it.endInclusive.toInt().toFloat()
                                     )
                                 )
                             )
